Add tutorial step7 - handling signals
In this version, we move the accept loop to a new coroutine, so main()
can wait on a signal pipe and exit cleanly when interrupted.

Clean exit is required for profilers (e.g. gprof) that use atexit() to
write the profile data. This function is not called if the process is
terminate in abnormal way.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>