Avoid core dumps in panic tests
Testing that child process panics generate unwanted and unhelpful core
dumps for successful tests. To avoid the core dumps these tests handle
now the SIGABRT and exit normally if libmill panics.

If libmill did not panic, the child process dump code so we can
investigate why it did not panic.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>